A Report for Notes as primary objects.
Driven by 	bug report 0013309: New Note Report  	
Able to use Note Filters.
Able to Sort the Notes.
Selectable languages for labels in the Notes but does not translate content.

Developer (George Baynes) does not have a GitHub account to get this published.
He is currently on 5.1.5 and the report GPR file is set up for the 5.1 major version. But it should only need the Gramps version changing in that file to work on other version.
Links work (http) but internal ones may not. Developer did not know what internal ones should do when written to a word processor document.
Developer does not use attributes... so they are untested.